Assignments/Hearth Failure/deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
plotly
Pillow
numpy
scikit-learn==1.2.2